requirements.txt
flask
tensorflow
numpy
pillow
gunicorn